chore: move fetchLayerFeaturesResource into the same file with layerFeaturesCollectionAtom
